Cleanup/tweak specs
